The server delegate protocol now includes the XMPPConnectionDelegate for element chain forwarding.
Added a guard when adding the connection to the registered connections dictionary; this will still work for P2P XMPP where a client can set the address during encapsulation before the final layer opens. And it will work for internet grade servers where the client authenticates/binds.

Added the sending of stream features to the XMPPConnection; they're sent even if empty - will need to compare this behaviour against the specification.

Moved the connection element forwarding into a separate class _XMPPForwarder so that implementors of the specific methods can duplicate the same forwarding behaviour.